Переезд интерфейсов и сервисов в отдельные проекты
