+++ Services/WebStore-Edu.Interfaces/Services/ICartService.cs
@@ -1,6 +1,6 @@
 ﻿using WebStore_Edu.Domain.ViewModels;
 
-namespace WebStore_Edu.Services.Interfaces
+namespace WebStore_Edu.Interfaces.Services
 {
     public interface ICartService
     {
