+++ Services/WebStore-Edu.Interfaces/Services/IProductData.cs
@@ -1,7 +1,7 @@
 ﻿using WebStore_Edu.Domain;
 using WebStore_Edu.Domain.Entityes;
 
-namespace WebStore_Edu.Services.Interfaces
+namespace WebStore_Edu.Interfaces.Services
 {
     public interface IProductData
     {
