+++ Services/WebStore-Edu.Services/Services/InSql/SqlEmployeesData.cs
@@ -1,10 +1,11 @@
 ﻿using MapsterMapper;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using WebStore_Edu.DAL.Context;
 using WebStore_Edu.Domain.Entityes;
-using WebStore_Edu.Services.Interfaces;
+using WebStore_Edu.Interfaces.Services;
 
-namespace WebStore_Edu.Services.InSql
+namespace WebStore_Edu.Services.Services.InSql
 {
     public class SqlEmployeesData : IEmployeesData
