+++ Services/WebStore-Edu.Services/Services/InSql/SqlOrderService.cs
@@ -5,9 +5,9 @@ using WebStore_Edu.DAL.Context;
 using WebStore_Edu.Domain.Entityes.Orders;
 using WebStore_Edu.Domain.Identity;
 using WebStore_Edu.Domain.ViewModels;
-using WebStore_Edu.Services.Interfaces;
+using WebStore_Edu.Interfaces.Services;
 
-namespace WebStore_Edu.Services.InSql
+namespace WebStore_Edu.Services.Services.InSql
 {
     public class SqlOrderService : IOrderService
     {
