+++ Services/WebStore-Edu.Services/Services/InSql/SqlProductData.cs
@@ -2,9 +2,9 @@
 using WebStore_Edu.DAL.Context;
 using WebStore_Edu.Domain;
 using WebStore_Edu.Domain.Entityes;
-using WebStore_Edu.Services.Interfaces;
+using WebStore_Edu.Interfaces.Services;
 
-namespace WebStore_Edu.Services.InSql
+namespace WebStore_Edu.Services.Services.InSql
 {
     public class SqlProductData : IProductData
     {
