+++ Services/WebStore-Edu.Services/Services/SqLiteDbInitializer.cs
@@ -1,20 +1,21 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using WebStore_Edu.DAL.Context;
-using WebStore_Edu.Data;
 using WebStore_Edu.Domain.Identity;
-using WebStore_Edu.Services.Interfaces;
+using WebStore_Edu.Interfaces.Services;
+using WebStore_Edu.Services.Data;
 
-namespace WebStore_Edu.Services
+namespace WebStore_Edu.Services.Services
 {
-    public class SqlServerDbInitializer : IDbInitializer
+    public class SqLiteDbInitializer: IDbInitializer
     {
         private readonly WebStoreDb _Db;
-        private readonly ILogger<SqlServerDbInitializer> _Logger;
+        private readonly ILogger<SqLiteDbInitializer> _Logger;
         private readonly UserManager<User> _UserManager;
         private readonly RoleManager<Role> _RoleManager;
 
-        public SqlServerDbInitializer(WebStoreDb Db, ILogger<SqlServerDbInitializer> Logger, UserManager<User> userManager, RoleManager<Role> roleManager)
+        public SqLiteDbInitializer(WebStoreDb Db, ILogger<SqLiteDbInitializer> Logger, UserManager<User> userManager, RoleManager<Role> roleManager)
         {
             _Db = Db;
             _Logger = Logger;
@@ -27,6 +28,8 @@ namespace WebStore_Edu.Services
             if (RemoveBefore) await RemoveAsync(Cancel).ConfigureAwait(false);
 
             _Logger.LogInformation("Инициализация БД...");
+
+            await _Db.Database.ExecuteSqlRawAsync("PRAGMA journal_mode=DELETE;", Cancel).ConfigureAwait(false);
 
             var pendingMigrations = await _Db.Database.GetPendingMigrationsAsync(Cancel).ConfigureAwait(false);
             if (pendingMigrations.Any())
@@ -69,9 +72,7 @@ namespace WebStore_Edu.Services
             await using (var transaction = await _Db.Database.BeginTransactionAsync(Cancel).ConfigureAwait(false))
             {
                 await _Db.Sections.AddRangeAsync(TestData.Sections, Cancel).ConfigureAwait(false);
-                await _Db.Database.ExecuteSqlRawAsync("SET IDENTITY_INSERT [dbo].[Sections] ON", Cancel); // TODO: плохо так делать
                 await _Db.SaveChangesAsync(Cancel).ConfigureAwait(false);
-                await _Db.Database.ExecuteSqlRawAsync("SET IDENTITY_INSERT [dbo].[Sections] OFF", Cancel);
                 await transaction.CommitAsync(Cancel).ConfigureAwait(false);
             }
             _Logger.LogInformation("Добавление категорий товаров выполнено");
@@ -81,9 +82,7 @@ namespace WebStore_Edu.Services
             await using (var transaction = await _Db.Database.BeginTransactionAsync(Cancel).ConfigureAwait(false))
             {
                 await _Db.Brands.AddRangeAsync(TestData.Brands, Cancel).ConfigureAwait(false);
-                await _Db.Database.ExecuteSqlRawAsync("SET IDENTITY_INSERT [dbo].[Brands] ON", Cancel);
                 await _Db.SaveChangesAsync(Cancel).ConfigureAwait(false);
-                await _Db.Database.ExecuteSqlRawAsync("SET IDENTITY_INSERT [dbo].[Brands] OFF", Cancel);
                 await transaction.CommitAsync(Cancel).ConfigureAwait(false);
             }
             _Logger.LogInformation("Добавление брендов выполнено");
@@ -93,9 +92,7 @@ namespace WebStore_Edu.Services
             await using (var transaction = await _Db.Database.BeginTransactionAsync(Cancel).ConfigureAwait(false))
             {
                 await _Db.Products.AddRangeAsync(TestData.Products, Cancel).ConfigureAwait(false);
-                await _Db.Database.ExecuteSqlRawAsync("SET IDENTITY_INSERT [dbo].[Products] ON", Cancel);
                 await _Db.SaveChangesAsync(Cancel).ConfigureAwait(false);
-                await _Db.Database.ExecuteSqlRawAsync("SET IDENTITY_INSERT [dbo].[Products] OFF", Cancel);
                 await transaction.CommitAsync(Cancel).ConfigureAwait(false);
             }
             _Logger.LogInformation("Добавление товаров выполнено");
