+++ Services/WebStore-Edu.Services/WebStore-Edu.Services.csproj
@@ -8,6 +8,11 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\..\Data\WebStore-Edu.DAL\WebStore-Edu.DAL.csproj" />
     <ProjectReference Include="..\WebStore-Edu.Interfaces\WebStore-Edu.Interfaces.csproj" />
   </ItemGroup>
