+++ UI/WebStore-Edu/Controllers/EmployeesController.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using WebStore_Edu.Domain.Entityes;
 using WebStore_Edu.Domain.Identity;
 using WebStore_Edu.Domain.ViewModels;
-using WebStore_Edu.Services.Interfaces;
+using WebStore_Edu.Interfaces.Services;
 
 namespace WebStore_Edu.Controllers
 {
