+++ UI/WebStore-Edu/Controllers/HomeController.cs
@@ -1,7 +1,7 @@
 ﻿using MapsterMapper;
 using Microsoft.AspNetCore.Mvc;
 using WebStore_Edu.Domain.ViewModels;
-using WebStore_Edu.Services.Interfaces;
+using WebStore_Edu.Interfaces.Services;
 
 namespace WebStore_Edu.Controllers
 {
