+++ UI/WebStore-Edu/Controllers/ShopController.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using WebStore_Edu.Domain;
 using WebStore_Edu.Domain.ViewModels;
-using WebStore_Edu.Services.Interfaces;
+using WebStore_Edu.Interfaces.Services;
 
 namespace WebStore_Edu.Controllers
 {
